Added Player Class
But lots of bugs introduced, did not create a Pokemon class which is causing issue
