+++ Pokemon/Utility.hpp
@@ -1,5 +1,3 @@
-#include <iostream>
-
 class Utility
 {
 public:
