+++ Pokemon/main.cpp
@@ -3,6 +3,7 @@
 #include <limits>
 #include "EPokemonChoice.hpp"
 #include "EPokemonType.hpp"
+#include "Player.hpp"
 #include "Utility.hpp"
 using namespace std;
 
@@ -44,49 +45,49 @@ public:
 	}
 };
 
-class Player
-{
-public:
-	string Name;
-	Pokemon ChosenPokemon;
-
-	Player() {
-		Name = "Trainer";
-		ChosenPokemon = Pokemon();
-	}
-
-	Player(std::string name, Pokemon chosenPokemon) 
-	{
-		Name = name;
-		ChosenPokemon = chosenPokemon;
-	}
-
-	Player(const Player& other)
-	{
-		Name = other.Name;
-		ChosenPokemon = other.ChosenPokemon;
-	}
-
-	void ChoosePokemon(int choice)
-	{
-		switch (choice)
-		{
-		case 1:
-			ChosenPokemon = Pokemon("Bulbasaur", PokemonType::Grass, 100);
-			break;
-		case 2:
-			ChosenPokemon = Pokemon("Charmander", PokemonType::Fire, 100);
-			break;
-		case 3:
-			ChosenPokemon = Pokemon("Squirtle", PokemonType::Water, 100);
-			break;
-		default:
-			ChosenPokemon = Pokemon("Pikachu", PokemonType::Electric, 100);
-			break;
-		}
-		cout << "System: You have chosen " << ChosenPokemon.Name << " as your first Pokemon!\n";
-	}
-};
+//class Player
+//{
+//public:
+//	string Name;
+//	Pokemon ChosenPokemon;
+//
+//	Player() {
+//		Name = "Trainer";
+//		ChosenPokemon = Pokemon();
+//	}
+//
+//	Player(std::string name, Pokemon chosenPokemon) 
+//	{
+//		Name = name;
+//		ChosenPokemon = chosenPokemon;
+//	}
+//
+//	Player(const Player& other)
+//	{
+//		Name = other.Name;
+//		ChosenPokemon = other.ChosenPokemon;
+//	}
+//
+//	void ChoosePokemon(int choice)
+//	{
+//		switch (choice)
+//		{
+//		case 1:
+//			ChosenPokemon = Pokemon("Bulbasaur", PokemonType::Grass, 100);
+//			break;
+//		case 2:
+//			ChosenPokemon = Pokemon("Charmander", PokemonType::Fire, 100);
+//			break;
+//		case 3:
+//			ChosenPokemon = Pokemon("Squirtle", PokemonType::Water, 100);
+//			break;
+//		default:
+//			ChosenPokemon = Pokemon("Pikachu", PokemonType::Electric, 100);
+//			break;
+//		}
+//		cout << "System: You have chosen " << ChosenPokemon.Name << " as your first Pokemon!\n";
+//	}
+//};
 
 class ProfessorOak
 {
